SNOMED CT Terminology Server REST API operations

Low-level wrapper functions based on httr for interfacing with the operations built in the Snowstorm API

api_concept(
  conceptId,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE
)

api_concepts(
  term = NULL,
  conceptIds = NULL,
  activeFilter = NULL,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  limit = snomedizer_options_get("limit"),
  offset = 0,
  catch404 = TRUE,
  ...
)

api_concept_descendants(
  conceptId,
  stated = FALSE,
  limit = snomedizer_options_get("limit"),
  offset = 0,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE
)

api_concept_descriptions(
  conceptId,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE
)

api_all_branches(
  endpoint = snomedizer_options_get("endpoint"),
  catch404 = TRUE
)

api_branch(
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE,
  ...
)

api_branch_descendants(
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE,
  ...
)

api_descriptions(
  concept,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  offset = 0,
  limit = snomedizer_options_get("limit"),
  catch404 = TRUE,
  ...
)

api_version(endpoint = snomedizer_options_get("endpoint"), catch404 = TRUE)

api_browser_concepts(
  conceptId,
  descendantCountForm = c(NULL, "inferred", "stated", "additional"),
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE
)

api_browser_concept_ancestors(
  conceptId,
  form = "inferred",
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE
)

api_browser_concept_children(
  conceptId,
  form = "inferred",
  includeDescendantCount = FALSE,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE
)

api_browser_concept_parents(
  conceptId,
  form = "inferred",
  includeDescendantCount = TRUE,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE
)

api_browser_concept_descriptions(
  term,
  active = NULL,
  module = NULL,
  language = NULL,
  type = NULL,
  semanticTag = NULL,
  semanticTags = NULL,
  preferredIn = NULL,
  acceptableIn = NULL,
  preferredOrAcceptableIn = NULL,
  conceptActive = NULL,
  conceptRefset = NULL,
  groupByConcept = FALSE,
  searchMode = "STANDARD",
  limit = snomedizer_options_get("limit"),
  offset = 0,
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE
)

api_descriptions_semantic_tags(
  endpoint = snomedizer_options_get("endpoint"),
  branch = snomedizer_options_get("branch"),
  catch404 = TRUE
)

Arguments

conceptId
character string of a SNOMED-CT concept id (for example: "233604007")

endpoint
the URL of a SNOMED CT Terminology Server REST API endpoint. See snomedizer_options.

branch
a string for the name of the API endpoint branch to use (most commonly "MAIN"). See snomedizer_options.

catch404
whether to display a warning if the API operation returns a '404 Not Found' status. Default is TRUE.

term
character vector of terms to search

conceptIds
a character vector of SNOMED-CT concept ids (for example: c("233604007", "68566005"))

activeFilter
optional boolean:

TRUE returns only active terminology

FALSE returns only inactive terminology

NULL (the default) returns both active and inactive terminology

limit
integer for the maximum number of results to return. See snomedizer_options.

offset
an integer indicating the number of results to skip

...
other REST API parameters

stated
a boolean indicating whether to limit search to descendants whose relationship is stated rather than inferred. Default is FALSE.

concept
character string of a SNOMED-CT concept id (for example: "233604007")

descendantCountForm
a character string indicating whether to report the count of descendant concepts based on stated or inferred relationships. Must be one of "inferred", "stated", or "additional". Default is NULL for no descendant count reported.

form
a character string indicating which ancestors/parents or descendants/children to extract based on stated or inferred relationships. Must be one of "inferred" (default), "stated", or "additional".

includeDescendantCount
a boolean indicating whether a number of children/descendants counter should be included in the result

active
optional boolean:

TRUE returns only active terminology

FALSE returns only inactive terminology

NULL (the default) returns both active and inactive terminology

module
character vector of SNOMED-CT modules to include (example: "900000000000207008")

language
vector of two-character language codes to include (example: c("en", "de")).

type
character vector of description types to include. See api_concept_descendants("900000000000446008") for valid description type inputs.

semanticTag
character string of a description semantic tag to include (example: "attribute"). See api_descriptions_semantic_tags() for a list of valid description semantic tags.

semanticTags
character vector of description semantic tags to include (example: c("attribute", "finding")). See api_descriptions_semantic_tags() for a list of valid description semantic tags.

preferredIn
character vector of description language reference sets (example: "900000000000509007"). The description must be preferred in at least one of these to match.

acceptableIn
character vector of description language reference sets (example: "900000000000509007"). The description must be acceptable in at least one of these to match.

preferredOrAcceptableIn
character vector of description language reference sets (example: "900000000000509007"). The description must be preferred OR acceptable in at least one of these to match.

conceptActive
optional boolean:

TRUE returns only active concepts

FALSE returns only inactive concepts

NULL (the default) returns both active and inactive concepts

conceptRefset
character vector of reference sets concept ids to include (example: "900000000000497000" for CTV3 terminology). See api_concept_descendants("900000000000455006") for valid reference set concepts.

groupByConcept
a boolean indicating whether to group descriptions by concept. Default is FALSE.

searchMode
a character string for the search mode. Must be either "STANDARD" (default) or "REGEX".

Value

An httr response() object.

Examples

# look up the pneumonia concept
api_concept(conceptId = "233604007")

#> Response [https://snowstorm.ihtsdotools.org/snowstorm/snomed-ct/MAIN/concepts/233604007]
#>   Date: 2020-07-19 19:36
#>   Status: 200
#>   Content-Type: application/json;charset=UTF-8
#>   Size: 313 B
#> {
#>   "conceptId" : "233604007",
#>   "active" : true,
#>   "definitionStatus" : "FULLY_DEFINED",
#>   "moduleId" : "900000000000207008",
#>   "effectiveTime" : "20150131",
#>   "fsn" : {
#>     "term" : "Pneumonia (disorder)",
#>     "lang" : "en"
#>   },
#> ...

api_concepts(term = "pneumonia")

#> Response [https://snowstorm.ihtsdotools.org/snowstorm/snomed-ct/MAIN/concepts?term=pneumonia&limit=50&offset=0]
#>   Date: 2020-07-19 19:36
#>   Status: 200
#>   Content-Type: application/json;charset=UTF-8
#>   Size: 18.4 kB
#> {
#>   "items" : [ {
#>     "conceptId" : "60363000",
#>     "active" : false,
#>     "definitionStatus" : "PRIMITIVE",
#>     "moduleId" : "900000000000207008",
#>     "effectiveTime" : "20020131",
#>     "fsn" : {
#>       "term" : "Pneumonia (disorder)",
#>       "lang" : "en"
#> ...

api_concepts(conceptIds = c("233604007", "68566005"))

#> Response [https://snowstorm.ihtsdotools.org/snowstorm/snomed-ct/MAIN/concepts?conceptIds=233604007&conceptIds=68566005&limit=50&offset=0]
#>   Date: 2020-07-19 19:36
#>   Status: 200
#>   Content-Type: application/json;charset=UTF-8
#>   Size: 875 B
#> {
#>   "items" : [ {
#>     "conceptId" : "68566005",
#>     "active" : true,
#>     "definitionStatus" : "FULLY_DEFINED",
#>     "moduleId" : "900000000000207008",
#>     "effectiveTime" : "20020131",
#>     "fsn" : {
#>       "term" : "Urinary tract infectious disease (disorder)",
#>       "lang" : "en"
#> ...

# get the content of the server request
httr::content(api_concepts(term = "pneumonia"))

#> $items
#> $items[[1]]
#> $items[[1]]$conceptId
#> [1] "60363000"
#> 
#> $items[[1]]$active
#> [1] FALSE
#> 
#> $items[[1]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[1]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[1]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[1]]$fsn
#> $items[[1]]$fsn$term
#> [1] "Pneumonia (disorder)"
#> 
#> $items[[1]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[1]]$pt
#> $items[[1]]$pt$term
#> [1] "Pneumonia"
#> 
#> $items[[1]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[1]]$id
#> [1] "60363000"
#> 
#> 
#> $items[[2]]
#> $items[[2]]$conceptId
#> [1] "266391003"
#> 
#> $items[[2]]$active
#> [1] FALSE
#> 
#> $items[[2]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[2]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[2]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[2]]$fsn
#> $items[[2]]$fsn$term
#> [1] "Pneumonia and influenza &/or pneumonia (disorder)"
#> 
#> $items[[2]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[2]]$pt
#> $items[[2]]$pt$term
#> [1] "Pneumonia and influenza &/or pneumonia"
#> 
#> $items[[2]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[2]]$id
#> [1] "266391003"
#> 
#> 
#> $items[[3]]
#> $items[[3]]$conceptId
#> [1] "233604007"
#> 
#> $items[[3]]$active
#> [1] TRUE
#> 
#> $items[[3]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[3]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[3]]$effectiveTime
#> [1] "20150131"
#> 
#> $items[[3]]$fsn
#> $items[[3]]$fsn$term
#> [1] "Pneumonia (disorder)"
#> 
#> $items[[3]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[3]]$pt
#> $items[[3]]$pt$term
#> [1] "Pneumonia"
#> 
#> $items[[3]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[3]]$id
#> [1] "233604007"
#> 
#> 
#> $items[[4]]
#> $items[[4]]$conceptId
#> [1] "155548002"
#> 
#> $items[[4]]$active
#> [1] FALSE
#> 
#> $items[[4]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[4]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[4]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[4]]$fsn
#> $items[[4]]$fsn$term
#> [1] "Pneumonia and influenza &/or pneumonia (disorder)"
#> 
#> $items[[4]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[4]]$pt
#> $items[[4]]$pt$term
#> [1] "Pneumonia and influenza &/or pneumonia"
#> 
#> $items[[4]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[4]]$id
#> [1] "155548002"
#> 
#> 
#> $items[[5]]
#> $items[[5]]$conceptId
#> [1] "155552002"
#> 
#> $items[[5]]$active
#> [1] FALSE
#> 
#> $items[[5]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[5]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[5]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[5]]$fsn
#> $items[[5]]$fsn$term
#> [1] "Pneumonia NOS (disorder)"
#> 
#> $items[[5]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[5]]$pt
#> $items[[5]]$pt$term
#> [1] "Pneumonia NOS"
#> 
#> $items[[5]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[5]]$id
#> [1] "155552002"
#> 
#> 
#> $items[[6]]
#> $items[[6]]$conceptId
#> [1] "155558003"
#> 
#> $items[[6]]$active
#> [1] FALSE
#> 
#> $items[[6]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[6]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[6]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[6]]$fsn
#> $items[[6]]$fsn$term
#> [1] "Pneumonia NOS (disorder)"
#> 
#> $items[[6]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[6]]$pt
#> $items[[6]]$pt$term
#> [1] "Pneumonia NOS"
#> 
#> $items[[6]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[6]]$id
#> [1] "155558003"
#> 
#> 
#> $items[[7]]
#> $items[[7]]$conceptId
#> [1] "274103002"
#> 
#> $items[[7]]$active
#> [1] FALSE
#> 
#> $items[[7]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[7]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[7]]$effectiveTime
#> [1] "20100131"
#> 
#> $items[[7]]$fsn
#> $items[[7]]$fsn$term
#> [1] "Pneumonia NOS (disorder)"
#> 
#> $items[[7]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[7]]$pt
#> $items[[7]]$pt$term
#> [1] "Pneumonia NOS"
#> 
#> $items[[7]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[7]]$id
#> [1] "274103002"
#> 
#> 
#> $items[[8]]
#> $items[[8]]$conceptId
#> [1] "138794008"
#> 
#> $items[[8]]$active
#> [1] FALSE
#> 
#> $items[[8]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[8]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[8]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[8]]$fsn
#> $items[[8]]$fsn$term
#> [1] "H/O: pneumonia (situation)"
#> 
#> $items[[8]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[8]]$pt
#> $items[[8]]$pt$term
#> [1] "H/O: pneumonia"
#> 
#> $items[[8]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[8]]$id
#> [1] "138794008"
#> 
#> 
#> $items[[9]]
#> $items[[9]]$conceptId
#> [1] "161525004"
#> 
#> $items[[9]]$active
#> [1] TRUE
#> 
#> $items[[9]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[9]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[9]]$effectiveTime
#> [1] "20040731"
#> 
#> $items[[9]]$fsn
#> $items[[9]]$fsn$term
#> [1] "History of pneumonia (situation)"
#> 
#> $items[[9]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[9]]$pt
#> $items[[9]]$pt$term
#> [1] "H/O: pneumonia"
#> 
#> $items[[9]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[9]]$id
#> [1] "161525004"
#> 
#> 
#> $items[[10]]
#> $items[[10]]$conceptId
#> [1] "233611006"
#> 
#> $items[[10]]$active
#> [1] FALSE
#> 
#> $items[[10]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[10]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[10]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[10]]$fsn
#> $items[[10]]$fsn$term
#> [1] "Fetal pneumonia (disorder)"
#> 
#> $items[[10]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[10]]$pt
#> $items[[10]]$pt$term
#> [1] "Fetal pneumonia"
#> 
#> $items[[10]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[10]]$id
#> [1] "233611006"
#> 
#> 
#> $items[[11]]
#> $items[[11]]$conceptId
#> [1] "233612004"
#> 
#> $items[[11]]$active
#> [1] FALSE
#> 
#> $items[[11]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[11]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[11]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[11]]$fsn
#> $items[[11]]$fsn$term
#> [1] "Fetal pneumonia (disorder)"
#> 
#> $items[[11]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[11]]$pt
#> $items[[11]]$pt$term
#> [1] "Fetal pneumonia"
#> 
#> $items[[11]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[11]]$id
#> [1] "233612004"
#> 
#> 
#> $items[[12]]
#> $items[[12]]$conceptId
#> [1] "155549005"
#> 
#> $items[[12]]$active
#> [1] FALSE
#> 
#> $items[[12]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[12]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[12]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[12]]$fsn
#> $items[[12]]$fsn$term
#> [1] "Viral pneumonia (disorder)"
#> 
#> $items[[12]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[12]]$pt
#> $items[[12]]$pt$term
#> [1] "Viral pneumonia"
#> 
#> $items[[12]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[12]]$id
#> [1] "155549005"
#> 
#> 
#> $items[[13]]
#> $items[[13]]$conceptId
#> [1] "37629004"
#> 
#> $items[[13]]$active
#> [1] FALSE
#> 
#> $items[[13]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[13]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[13]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[13]]$fsn
#> $items[[13]]$fsn$term
#> [1] "Pneumonitis due to inhalation of oils (disorder)"
#> 
#> $items[[13]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[13]]$pt
#> $items[[13]]$pt$term
#> [1] "Pneumonitis due to inhalation of oils"
#> 
#> $items[[13]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[13]]$id
#> [1] "37629004"
#> 
#> 
#> $items[[14]]
#> $items[[14]]$conceptId
#> [1] "416916004"
#> 
#> $items[[14]]$active
#> [1] TRUE
#> 
#> $items[[14]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[14]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[14]]$effectiveTime
#> [1] "20050731"
#> 
#> $items[[14]]$fsn
#> $items[[14]]$fsn$term
#> [1] "Lipoid pneumonitis (disorder)"
#> 
#> $items[[14]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[14]]$pt
#> $items[[14]]$pt$term
#> [1] "Lipoid pneumonitis"
#> 
#> $items[[14]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[14]]$id
#> [1] "416916004"
#> 
#> 
#> $items[[15]]
#> $items[[15]]$conceptId
#> [1] "300999006"
#> 
#> $items[[15]]$active
#> [1] TRUE
#> 
#> $items[[15]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[15]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[15]]$effectiveTime
#> [1] "20160731"
#> 
#> $items[[15]]$fsn
#> $items[[15]]$fsn$term
#> [1] "Basal pneumonia (disorder)"
#> 
#> $items[[15]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[15]]$pt
#> $items[[15]]$pt$term
#> [1] "Basal pneumonia"
#> 
#> $items[[15]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[15]]$id
#> [1] "300999006"
#> 
#> 
#> $items[[16]]
#> $items[[16]]$conceptId
#> [1] "278516003"
#> 
#> $items[[16]]$active
#> [1] TRUE
#> 
#> $items[[16]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[16]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[16]]$effectiveTime
#> [1] "20150731"
#> 
#> $items[[16]]$fsn
#> $items[[16]]$fsn$term
#> [1] "Lobar pneumonia (disorder)"
#> 
#> $items[[16]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[16]]$pt
#> $items[[16]]$pt$term
#> [1] "Lobar pneumonia"
#> 
#> $items[[16]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[16]]$id
#> [1] "278516003"
#> 
#> 
#> $items[[17]]
#> $items[[17]]$conceptId
#> [1] "195880002"
#> 
#> $items[[17]]$active
#> [1] FALSE
#> 
#> $items[[17]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[17]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[17]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[17]]$fsn
#> $items[[17]]$fsn$term
#> [1] "Viral pneumonia (& chest infection) (disorder)"
#> 
#> $items[[17]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[17]]$pt
#> $items[[17]]$pt$term
#> [1] "Viral pneumonia (& chest infection)"
#> 
#> $items[[17]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[17]]$id
#> [1] "195880002"
#> 
#> 
#> $items[[18]]
#> $items[[18]]$conceptId
#> [1] "75570004"
#> 
#> $items[[18]]$active
#> [1] TRUE
#> 
#> $items[[18]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[18]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[18]]$effectiveTime
#> [1] "20130731"
#> 
#> $items[[18]]$fsn
#> $items[[18]]$fsn$term
#> [1] "Viral pneumonia (disorder)"
#> 
#> $items[[18]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[18]]$pt
#> $items[[18]]$pt$term
#> [1] "Viral pneumonia"
#> 
#> $items[[18]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[18]]$id
#> [1] "75570004"
#> 
#> 
#> $items[[19]]
#> $items[[19]]$conceptId
#> [1] "266392005"
#> 
#> $items[[19]]$active
#> [1] FALSE
#> 
#> $items[[19]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[19]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[19]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[19]]$fsn
#> $items[[19]]$fsn$term
#> [1] "Pneumonia: [lobar] or [pneumococcal] (disorder)"
#> 
#> $items[[19]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[19]]$pt
#> $items[[19]]$pt$term
#> [1] "Pneumonia: [lobar] or [pneumococcal]"
#> 
#> $items[[19]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[19]]$id
#> [1] "266392005"
#> 
#> 
#> $items[[20]]
#> $items[[20]]$conceptId
#> [1] "123590007"
#> 
#> $items[[20]]$active
#> [1] TRUE
#> 
#> $items[[20]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[20]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[20]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[20]]$fsn
#> $items[[20]]$fsn$term
#> [1] "Focal pneumonia (disorder)"
#> 
#> $items[[20]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[20]]$pt
#> $items[[20]]$pt$term
#> [1] "Focal pneumonia"
#> 
#> $items[[20]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[20]]$id
#> [1] "123590007"
#> 
#> 
#> $items[[21]]
#> $items[[21]]$conceptId
#> [1] "54339004"
#> 
#> $items[[21]]$active
#> [1] FALSE
#> 
#> $items[[21]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[21]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[21]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[21]]$fsn
#> $items[[21]]$fsn$term
#> [1] "Pneumococcal pneumonia (disorder)"
#> 
#> $items[[21]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[21]]$pt
#> $items[[21]]$pt$term
#> [1] "Pneumococcal pneumonia"
#> 
#> $items[[21]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[21]]$id
#> [1] "54339004"
#> 
#> 
#> $items[[22]]
#> $items[[22]]$conceptId
#> [1] "196038008"
#> 
#> $items[[22]]$active
#> [1] FALSE
#> 
#> $items[[22]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[22]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[22]]$effectiveTime
#> [1] "20050731"
#> 
#> $items[[22]]$fsn
#> $items[[22]]$fsn$term
#> [1] "Lipoid pneumonitis (disorder)"
#> 
#> $items[[22]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[22]]$pt
#> $items[[22]]$pt$term
#> [1] "Pneumonitis due to inhalation of oils"
#> 
#> $items[[22]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[22]]$id
#> [1] "196038008"
#> 
#> 
#> $items[[23]]
#> $items[[23]]$conceptId
#> [1] "233730002"
#> 
#> $items[[23]]$active
#> [1] TRUE
#> 
#> $items[[23]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[23]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[23]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[23]]$fsn
#> $items[[23]]$fsn$term
#> [1] "Lupus pneumonia (disorder)"
#> 
#> $items[[23]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[23]]$pt
#> $items[[23]]$pt$term
#> [1] "Lupus pneumonia"
#> 
#> $items[[23]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[23]]$id
#> [1] "233730002"
#> 
#> 
#> $items[[24]]
#> $items[[24]]$conceptId
#> [1] "155550005"
#> 
#> $items[[24]]$active
#> [1] FALSE
#> 
#> $items[[24]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[24]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[24]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[24]]$fsn
#> $items[[24]]$fsn$term
#> [1] "Pneumonia: [lobar] or [pneumococcal] (disorder)"
#> 
#> $items[[24]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[24]]$pt
#> $items[[24]]$pt$term
#> [1] "Pneumonia: [lobar] or [pneumococcal]"
#> 
#> $items[[24]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[24]]$id
#> [1] "155550005"
#> 
#> 
#> $items[[25]]
#> $items[[25]]$conceptId
#> [1] "71926009"
#> 
#> $items[[25]]$active
#> [1] TRUE
#> 
#> $items[[25]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[25]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[25]]$effectiveTime
#> [1] "20150731"
#> 
#> $items[[25]]$fsn
#> $items[[25]]$fsn$term
#> [1] "Infective pneumonia acquired prenatally (disorder)"
#> 
#> $items[[25]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[25]]$pt
#> $items[[25]]$pt$term
#> [1] "Infective pneumonia acquired prenatally"
#> 
#> $items[[25]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[25]]$id
#> [1] "71926009"
#> 
#> 
#> $items[[26]]
#> $items[[26]]$conceptId
#> [1] "233613009"
#> 
#> $items[[26]]$active
#> [1] TRUE
#> 
#> $items[[26]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[26]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[26]]$effectiveTime
#> [1] "20190131"
#> 
#> $items[[26]]$fsn
#> $items[[26]]$fsn$term
#> [1] "Fungal pneumonia (disorder)"
#> 
#> $items[[26]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[26]]$pt
#> $items[[26]]$pt$term
#> [1] "Fungal pneumonia"
#> 
#> $items[[26]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[26]]$id
#> [1] "233613009"
#> 
#> 
#> $items[[27]]
#> $items[[27]]$conceptId
#> [1] "123595002"
#> 
#> $items[[27]]$active
#> [1] TRUE
#> 
#> $items[[27]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[27]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[27]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[27]]$fsn
#> $items[[27]]$fsn$term
#> [1] "Uremic lung (disorder)"
#> 
#> $items[[27]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[27]]$pt
#> $items[[27]]$pt$term
#> [1] "Uremic lung"
#> 
#> $items[[27]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[27]]$id
#> [1] "123595002"
#> 
#> 
#> $items[[28]]
#> $items[[28]]$conceptId
#> [1] "195888009"
#> 
#> $items[[28]]$active
#> [1] TRUE
#> 
#> $items[[28]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[28]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[28]]$effectiveTime
#> [1] "20020731"
#> 
#> $items[[28]]$fsn
#> $items[[28]]$fsn$term
#> [1] "Proteus pneumonia (disorder)"
#> 
#> $items[[28]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[28]]$pt
#> $items[[28]]$pt$term
#> [1] "Proteus pneumonia"
#> 
#> $items[[28]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[28]]$id
#> [1] "195888009"
#> 
#> 
#> $items[[29]]
#> $items[[29]]$conceptId
#> [1] "155555000"
#> 
#> $items[[29]]$active
#> [1] FALSE
#> 
#> $items[[29]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[29]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[29]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[29]]$fsn
#> $items[[29]]$fsn$term
#> [1] "Measles pneumonia (disorder)"
#> 
#> $items[[29]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[29]]$pt
#> $items[[29]]$pt$term
#> [1] "Measles pneumonia"
#> 
#> $items[[29]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[29]]$id
#> [1] "155555000"
#> 
#> 
#> $items[[30]]
#> $items[[30]]$conceptId
#> [1] "32286006"
#> 
#> $items[[30]]$active
#> [1] TRUE
#> 
#> $items[[30]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[30]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[30]]$effectiveTime
#> [1] "20160131"
#> 
#> $items[[30]]$fsn
#> $items[[30]]$fsn$term
#> [1] "Pneumonia in Q fever (disorder)"
#> 
#> $items[[30]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[30]]$pt
#> $items[[30]]$pt$term
#> [1] "Pneumonia in Q fever"
#> 
#> $items[[30]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[30]]$id
#> [1] "32286006"
#> 
#> 
#> $items[[31]]
#> $items[[31]]$conceptId
#> [1] "45312009"
#> 
#> $items[[31]]$active
#> [1] TRUE
#> 
#> $items[[31]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[31]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[31]]$effectiveTime
#> [1] "20160131"
#> 
#> $items[[31]]$fsn
#> $items[[31]]$fsn$term
#> [1] "Pneumonia in typhoid fever (disorder)"
#> 
#> $items[[31]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[31]]$pt
#> $items[[31]]$pt$term
#> [1] "Pneumonia in typhoid fever"
#> 
#> $items[[31]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[31]]$id
#> [1] "45312009"
#> 
#> 
#> $items[[32]]
#> $items[[32]]$conceptId
#> [1] "46970008"
#> 
#> $items[[32]]$active
#> [1] TRUE
#> 
#> $items[[32]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[32]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[32]]$effectiveTime
#> [1] "20100131"
#> 
#> $items[[32]]$fsn
#> $items[[32]]$fsn$term
#> [1] "Pneumonia caused by Mycoplasma pneumoniae (disorder)"
#> 
#> $items[[32]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[32]]$pt
#> $items[[32]]$pt$term
#> [1] "Mycoplasma pneumonia"
#> 
#> $items[[32]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[32]]$id
#> [1] "46970008"
#> 
#> 
#> $items[[33]]
#> $items[[33]]$conceptId
#> [1] "195900001"
#> 
#> $items[[33]]$active
#> [1] TRUE
#> 
#> $items[[33]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[33]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[33]]$effectiveTime
#> [1] "20020731"
#> 
#> $items[[33]]$fsn
#> $items[[33]]$fsn$term
#> [1] "Pneumonia co-occurrent and due to measles (disorder)"
#> 
#> $items[[33]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[33]]$pt
#> $items[[33]]$pt$term
#> [1] "Measles pneumonia"
#> 
#> $items[[33]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[33]]$id
#> [1] "195900001"
#> 
#> 
#> $items[[34]]
#> $items[[34]]$conceptId
#> [1] "102361000119104"
#> 
#> $items[[34]]$active
#> [1] TRUE
#> 
#> $items[[34]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[34]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[34]]$effectiveTime
#> [1] "20150731"
#> 
#> $items[[34]]$fsn
#> $items[[34]]$fsn$term
#> [1] "Chronic pneumonia (disorder)"
#> 
#> $items[[34]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[34]]$pt
#> $items[[34]]$pt$term
#> [1] "Chronic pneumonia"
#> 
#> $items[[34]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[34]]$id
#> [1] "102361000119104"
#> 
#> 
#> $items[[35]]
#> $items[[35]]$conceptId
#> [1] "85469005"
#> 
#> $items[[35]]$active
#> [1] TRUE
#> 
#> $items[[35]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[35]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[35]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[35]]$fsn
#> $items[[35]]$fsn$term
#> [1] "Hypostatic pneumonia (disorder)"
#> 
#> $items[[35]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[35]]$pt
#> $items[[35]]$pt$term
#> [1] "Hypostatic pneumonia"
#> 
#> $items[[35]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[35]]$id
#> [1] "85469005"
#> 
#> 
#> $items[[36]]
#> $items[[36]]$conceptId
#> [1] "57716000"
#> 
#> $items[[36]]$active
#> [1] FALSE
#> 
#> $items[[36]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[36]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[36]]$effectiveTime
#> [1] "20140131"
#> 
#> $items[[36]]$fsn
#> $items[[36]]$fsn$term
#> [1] "Brooder pneumonia (disorder)"
#> 
#> $items[[36]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[36]]$pt
#> $items[[36]]$pt$term
#> [1] "Brooder pneumonia"
#> 
#> $items[[36]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[36]]$id
#> [1] "57716000"
#> 
#> 
#> $items[[37]]
#> $items[[37]]$conceptId
#> [1] "396284006"
#> 
#> $items[[37]]$active
#> [1] FALSE
#> 
#> $items[[37]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[37]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[37]]$effectiveTime
#> [1] "20170731"
#> 
#> $items[[37]]$fsn
#> $items[[37]]$fsn$term
#> [1] "Lobular pneumonia (disorder)"
#> 
#> $items[[37]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[37]]$pt
#> $items[[37]]$pt$term
#> [1] "Lobular pneumonia"
#> 
#> $items[[37]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[37]]$id
#> [1] "396284006"
#> 
#> 
#> $items[[38]]
#> $items[[38]]$conceptId
#> [1] "396285007"
#> 
#> $items[[38]]$active
#> [1] TRUE
#> 
#> $items[[38]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[38]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[38]]$effectiveTime
#> [1] "20150131"
#> 
#> $items[[38]]$fsn
#> $items[[38]]$fsn$term
#> [1] "Bronchopneumonia (disorder)"
#> 
#> $items[[38]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[38]]$pt
#> $items[[38]]$pt$term
#> [1] "Bronchopneumonia"
#> 
#> $items[[38]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[38]]$id
#> [1] "396285007"
#> 
#> 
#> $items[[39]]
#> $items[[39]]$conceptId
#> [1] "1092361000119109"
#> 
#> $items[[39]]$active
#> [1] TRUE
#> 
#> $items[[39]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[39]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[39]]$effectiveTime
#> [1] "20150131"
#> 
#> $items[[39]]$fsn
#> $items[[39]]$fsn$term
#> [1] "Rubella pneumonia (disorder)"
#> 
#> $items[[39]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[39]]$pt
#> $items[[39]]$pt$term
#> [1] "Rubella pneumonia"
#> 
#> $items[[39]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[39]]$id
#> [1] "1092361000119109"
#> 
#> 
#> $items[[40]]
#> $items[[40]]$conceptId
#> [1] "195902009"
#> 
#> $items[[40]]$active
#> [1] TRUE
#> 
#> $items[[40]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[40]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[40]]$effectiveTime
#> [1] "20150731"
#> 
#> $items[[40]]$fsn
#> $items[[40]]$fsn$term
#> [1] "Anthrax pneumonia (disorder)"
#> 
#> $items[[40]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[40]]$pt
#> $items[[40]]$pt$term
#> [1] "Anthrax pneumonia"
#> 
#> $items[[40]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[40]]$id
#> [1] "195902009"
#> 
#> 
#> $items[[41]]
#> $items[[41]]$conceptId
#> [1] "67814005"
#> 
#> $items[[41]]$active
#> [1] FALSE
#> 
#> $items[[41]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[41]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[41]]$effectiveTime
#> [1] "20030731"
#> 
#> $items[[41]]$fsn
#> $items[[41]]$fsn$term
#> [1] "Bronchopneumonia (disorder)"
#> 
#> $items[[41]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[41]]$pt
#> $items[[41]]$pt$term
#> [1] "Bronchopneumonia"
#> 
#> $items[[41]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[41]]$id
#> [1] "67814005"
#> 
#> 
#> $items[[42]]
#> $items[[42]]$conceptId
#> [1] "35037009"
#> 
#> $items[[42]]$active
#> [1] TRUE
#> 
#> $items[[42]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[42]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[42]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[42]]$fsn
#> $items[[42]]$fsn$term
#> [1] "Primary atypical interstitial pneumonia (disorder)"
#> 
#> $items[[42]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[42]]$pt
#> $items[[42]]$pt$term
#> [1] "Primary atypical interstitial pneumonia"
#> 
#> $items[[42]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[42]]$id
#> [1] "35037009"
#> 
#> 
#> $items[[43]]
#> $items[[43]]$conceptId
#> [1] "187019005"
#> 
#> $items[[43]]$active
#> [1] FALSE
#> 
#> $items[[43]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[43]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[43]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[43]]$fsn
#> $items[[43]]$fsn$term
#> [1] "Candidiasis: [lung] or [pneumonia] (disorder)"
#> 
#> $items[[43]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[43]]$pt
#> $items[[43]]$pt$term
#> [1] "Candidiasis: [lung] or [pneumonia]"
#> 
#> $items[[43]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[43]]$id
#> [1] "187019005"
#> 
#> 
#> $items[[44]]
#> $items[[44]]$conceptId
#> [1] "471272001"
#> 
#> $items[[44]]$active
#> [1] TRUE
#> 
#> $items[[44]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[44]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[44]]$effectiveTime
#> [1] "20200731"
#> 
#> $items[[44]]$fsn
#> $items[[44]]$fsn$term
#> [1] "Cavitary pneumonia (disorder)"
#> 
#> $items[[44]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[44]]$pt
#> $items[[44]]$pt$term
#> [1] "Cavitary pneumonia"
#> 
#> $items[[44]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[44]]$id
#> [1] "471272001"
#> 
#> 
#> $items[[45]]
#> $items[[45]]$conceptId
#> [1] "233606009"
#> 
#> $items[[45]]$active
#> [1] TRUE
#> 
#> $items[[45]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[45]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[45]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[45]]$fsn
#> $items[[45]]$fsn$term
#> [1] "Atypical pneumonia (disorder)"
#> 
#> $items[[45]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[45]]$pt
#> $items[[45]]$pt$term
#> [1] "Atypical pneumonia"
#> 
#> $items[[45]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[45]]$id
#> [1] "233606009"
#> 
#> 
#> $items[[46]]
#> $items[[46]]$conceptId
#> [1] "233619008"
#> 
#> $items[[46]]$active
#> [1] TRUE
#> 
#> $items[[46]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[46]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[46]]$effectiveTime
#> [1] "20150731"
#> 
#> $items[[46]]$fsn
#> $items[[46]]$fsn$term
#> [1] "Neonatal pneumonia (disorder)"
#> 
#> $items[[46]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[46]]$pt
#> $items[[46]]$pt$term
#> [1] "Neonatal pneumonia"
#> 
#> $items[[46]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[46]]$id
#> [1] "233619008"
#> 
#> 
#> $items[[47]]
#> $items[[47]]$conceptId
#> [1] "426696003"
#> 
#> $items[[47]]$active
#> [1] TRUE
#> 
#> $items[[47]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[47]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[47]]$fsn
#> $items[[47]]$fsn$term
#> [1] "Lingular pneumonia (disorder)"
#> 
#> $items[[47]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[47]]$pt
#> $items[[47]]$pt$term
#> [1] "Lingular pneumonia"
#> 
#> $items[[47]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[47]]$id
#> [1] "426696003"
#> 
#> 
#> $items[[48]]
#> $items[[48]]$conceptId
#> [1] "195932001"
#> 
#> $items[[48]]$active
#> [1] FALSE
#> 
#> $items[[48]]$definitionStatus
#> [1] "PRIMITIVE"
#> 
#> $items[[48]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[48]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[48]]$fsn
#> $items[[48]]$fsn$term
#> [1] "Atypical pneumonia (disorder)"
#> 
#> $items[[48]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[48]]$pt
#> $items[[48]]$pt$term
#> [1] "Atypical pneumonia"
#> 
#> $items[[48]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[48]]$id
#> [1] "195932001"
#> 
#> 
#> $items[[49]]
#> $items[[49]]$conceptId
#> [1] "88687001"
#> 
#> $items[[49]]$active
#> [1] TRUE
#> 
#> $items[[49]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[49]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[49]]$effectiveTime
#> [1] "20200731"
#> 
#> $items[[49]]$fsn
#> $items[[49]]$fsn$term
#> [1] "Manganese pneumonitis (disorder)"
#> 
#> $items[[49]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[49]]$pt
#> $items[[49]]$pt$term
#> [1] "Manganese pneumonitis"
#> 
#> $items[[49]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[49]]$id
#> [1] "88687001"
#> 
#> 
#> $items[[50]]
#> $items[[50]]$conceptId
#> [1] "3487004"
#> 
#> $items[[50]]$active
#> [1] TRUE
#> 
#> $items[[50]]$definitionStatus
#> [1] "FULLY_DEFINED"
#> 
#> $items[[50]]$moduleId
#> [1] "900000000000207008"
#> 
#> $items[[50]]$effectiveTime
#> [1] "20020131"
#> 
#> $items[[50]]$fsn
#> $items[[50]]$fsn$term
#> [1] "Candidiasis of lung (disorder)"
#> 
#> $items[[50]]$fsn$lang
#> [1] "en"
#> 
#> 
#> $items[[50]]$pt
#> $items[[50]]$pt$term
#> [1] "Candidiasis of lung"
#> 
#> $items[[50]]$pt$lang
#> [1] "en"
#> 
#> 
#> $items[[50]]$id
#> [1] "3487004"
#> 
#> 
#> 
#> $total
#> [1] 769
#> 
#> $limit
#> [1] 50
#> 
#> $offset
#> [1] 0
#> 
#> $searchAfter
#> [1] "WzM0ODcwMDRd"
#> 
#> $searchAfterArray
#> $searchAfterArray[[1]]
#> [1] 3487004
#> 
#>