Find SNOMED-CT concepts

A wrapper function searching for a term or a list of concepts

concepts_find(
  term = NULL,
  conceptIds = NULL,
  ecl = NULL,
  activeFilter = TRUE,
  silent = FALSE,
  ...
)

Arguments

term
a string containing search terms

conceptIds
a character vector of one or more SNOMED-CT codes to search

ecl
a character expression contraint. Consult the Expression Constraint Language guide for more detail.

activeFilter
whether to restrict results to active concepts. Default is `TRUE`. Consult the SNOMED glossary for more detail.

silent
whether to hide warnings. Default is `FALSE`

...
other optional arguments listed in api_operations

Value

a data frame

Examples

# Free text search
concepts_find("asthma")

#> Warning: 
#> This server request returned just 50 of a total 252 results.
#> Please increase the server `limit` to fetch all results.

#>    conceptId active definitionStatus           moduleId effectiveTime        id
#> 1  195967001   TRUE        PRIMITIVE 900000000000207008      20040731 195967001
#> 2  160377001   TRUE    FULLY_DEFINED 900000000000207008      20040731 160377001
#> 3   56968009   TRUE    FULLY_DEFINED 900000000000207008      20020731  56968009
#> 4  406162001   TRUE        PRIMITIVE 900000000000207008      20040131 406162001
#> 5  161527007   TRUE    FULLY_DEFINED 900000000000207008      20040731 161527007
#> 6  370218001   TRUE        PRIMITIVE 900000000000207008      20020731 370218001
#> 7  304527002   TRUE    FULLY_DEFINED 900000000000207008      20150731 304527002
#> 8  195977004   TRUE        PRIMITIVE 900000000000207008      20020131 195977004
#> 9   34015007   TRUE    FULLY_DEFINED 900000000000207008      20060131  34015007
#> 10 370221004   TRUE        PRIMITIVE 900000000000207008      20020731 370221004
#> 11 708038006   TRUE    FULLY_DEFINED 900000000000207008      20150731 708038006
#> 12 233683003   TRUE    FULLY_DEFINED 900000000000207008      20060131 233683003
#> 13 702823004   TRUE        PRIMITIVE 900000000000207008      20140731 702823004
#> 14 424643009   TRUE    FULLY_DEFINED 900000000000207008      20070131 424643009
#> 15 389145006   TRUE    FULLY_DEFINED 900000000000207008      20190131 389145006
#> 16 390798007   TRUE        PRIMITIVE 900000000000207008      20030131 390798007
#> 17 225057002   TRUE        PRIMITIVE 900000000000207008      20020131 225057002
#> 18 400987003   TRUE        PRIMITIVE 900000000000207008      20030731 400987003
#> 19  71892000   TRUE        PRIMITIVE 900000000000207008      20020131  71892000
#> 20  29422001   TRUE    FULLY_DEFINED 900000000000207008      20190731  29422001
#> 21  11944003   TRUE        PRIMITIVE 900000000000207008      20020131  11944003
#> 22  11641008   TRUE        PRIMITIVE 900000000000207008      20020131  11641008
#> 23 370219009   TRUE        PRIMITIVE 900000000000207008      20020731 370219009
#> 24 162660004   TRUE        PRIMITIVE 900000000000207008      20020131 162660004
#> 25 404804003   TRUE        PRIMITIVE 900000000000207008      20040131 404804003
#> 26 445427006   TRUE    FULLY_DEFINED 900000000000207008      20100731 445427006
#> 27 423889005   TRUE    FULLY_DEFINED 900000000000207008      20070131 423889005
#> 28 171231001   TRUE    FULLY_DEFINED 900000000000207008      20020131 171231001
#> 29 401193004   TRUE    FULLY_DEFINED 900000000000207008      20040731 401193004
#> 30 394701000   TRUE        PRIMITIVE 900000000000207008      20030131 394701000
#> 31 394967008   TRUE    FULLY_DEFINED 900000000000207008      20050731 394967008
#> 32  23315001   TRUE        PRIMITIVE 900000000000207008      20020131  23315001
#> 33 233687002   TRUE        PRIMITIVE 900000000000207008      20190131 233687002
#> 34 270442000   TRUE        PRIMITIVE 900000000000207008      20020131 270442000
#> 35 401135008   TRUE    FULLY_DEFINED 900000000000207008      20100731 401135008
#> 36 266361008   TRUE        PRIMITIVE 900000000000207008      20020131 266361008
#> 37 395022009   TRUE        PRIMITIVE 900000000000207008      20030131 395022009
#> 38 233678006   TRUE        PRIMITIVE 900000000000207008      20020131 233678006
#> 39  41553006   TRUE    FULLY_DEFINED 900000000000207008      20060131  41553006
#> 40  18041002   TRUE        PRIMITIVE 900000000000207008      20020131  18041002
#> 41  18197001   TRUE        PRIMITIVE 900000000000207008      20020131  18197001
#> 42 786836003   TRUE        PRIMITIVE 900000000000207008      20190731 786836003
#> 43  57607007   TRUE        PRIMITIVE 900000000000207008      20090731  57607007
#> 44 370220003   TRUE        PRIMITIVE 900000000000207008      20020731 370220003
#> 45 233679003   TRUE        PRIMITIVE 900000000000207008      20020131 233679003
#> 46 233690008   TRUE        PRIMITIVE 900000000000207008      20020131 233690008
#> 47 275908000   TRUE        PRIMITIVE 900000000000207008      20020131 275908000
#> 48  13151001   TRUE    FULLY_DEFINED 900000000000207008      20200131  13151001
#> 49  85761009   TRUE    FULLY_DEFINED 900000000000207008      20200131  85761009
#> 50 708090002   TRUE        PRIMITIVE 900000000000207008      20150131 708090002
#>                                                    fsn.term fsn.lang
#> 1                                         Asthma (disorder)       en
#> 2                        Family history: Asthma (situation)       en
#> 3                     Asthma caused by wood dust (disorder)       en
#> 4                        Asthma management (regime/therapy)       en
#> 5                             History of asthma (situation)       en
#> 6                                    Mild asthma (disorder)       en
#> 7                                   Acute asthma (disorder)       en
#> 8                                   Mixed asthma (disorder)       en
#> 9                                 Bakers' asthma (disorder)       en
#> 10                                 Severe asthma (disorder)       en
#> 11                  Acute exacerbation of asthma (disorder)       en
#> 12                         Hay fever with asthma (disorder)       en
#> 13                              Asthma clinic (environment)       en
#> 14     Immunoglobulin E-mediated allergic asthma (disorder)       en
#> 15                               Allergic asthma (disorder)       en
#> 16                                 Asthma finding (finding)       en
#> 17                                Brittle asthma (disorder)       en
#> 18                       Asthma trigger (observable entity)       en
#> 19                                Cardiac asthma (disorder)       en
#> 20                  Coal workers' pneumoconiosis (disorder)       en
#> 21                      Feather-pickers' disease (disorder)       en
#> 22                               Millers' asthma (disorder)       en
#> 23                               Moderate asthma (disorder)       en
#> 24                                Asthma resolved (finding)       en
#> 25                               Platinum asthma (disorder)       en
#> 26                               Seasonal asthma (disorder)       en
#> 27 Non-immunoglobulin E mediated allergic asthma (disorder)       en
#> 28                             Asthma screening (procedure)       en
#> 29                             Asthma confirmed (situation)       en
#> 30                        Asthma follow-up (regime/therapy)       en
#> 31                             Suspected asthma (situation)       en
#> 32                                    Sequoiosis (disorder)       en
#> 33                              Colophony asthma (disorder)       en
#> 34                   Asthma monitoring check done (finding)       en
#> 35                             Asthma education (procedure)       en
#> 36                           Non-allergic asthma (disorder)       en
#> 37                     Asthma night-time symptoms (finding)       en
#> 38                              Childhood asthma (disorder)       en
#> 39                              Detergent asthma (disorder)       en
#> 40                              Printers' asthma (disorder)       en
#> 41                              Asthmatoid wheeze (finding)       en
#> 42                             Near fatal asthma (disorder)       en
#> 43                           Occupational asthma (disorder)       en
#> 44                             Occasional asthma (disorder)       en
#> 45                             Late onset asthma (disorder)       en
#> 46                             Factitious asthma (disorder)       en
#> 47                       Asthma monitoring (regime/therapy)       en
#> 48                        Flax-dressers' disease (disorder)       en
#> 49                                    Byssinosis (disorder)       en
#> 50           Acute severe exacerbation of asthma (disorder)       en
#>                                pt.term pt.lang total limit offset
#> 1                               Asthma      en   252    50      0
#> 2             Family history of asthma      en   252    50      0
#> 3           Asthma caused by wood dust      en   252    50      0
#> 4                    Asthma management      en   252    50      0
#> 5                          H/O: asthma      en   252    50      0
#> 6                          Mild asthma      en   252    50      0
#> 7                         Acute asthma      en   252    50      0
#> 8                         Mixed asthma      en   252    50      0
#> 9                       Bakers' asthma      en   252    50      0
#> 10                       Severe asthma      en   252    50      0
#> 11        Acute exacerbation of asthma      en   252    50      0
#> 12               Hay fever with asthma      en   252    50      0
#> 13                       Asthma clinic      en   252    50      0
#> 14        IgE-mediated allergic asthma      en   252    50      0
#> 15                     Allergic asthma      en   252    50      0
#> 16                      Asthma finding      en   252    50      0
#> 17                      Brittle asthma      en   252    50      0
#> 18                      Asthma trigger      en   252    50      0
#> 19                      Cardiac asthma      en   252    50      0
#> 20        Coal workers' pneumoconiosis      en   252    50      0
#> 21            Feather-pickers' disease      en   252    50      0
#> 22                     Millers' asthma      en   252    50      0
#> 23                     Moderate asthma      en   252    50      0
#> 24                     Asthma resolved      en   252    50      0
#> 25                     Platinum asthma      en   252    50      0
#> 26                     Seasonal asthma      en   252    50      0
#> 27    Non-IgE mediated allergic asthma      en   252    50      0
#> 28                    Asthma screening      en   252    50      0
#> 29                    Asthma confirmed      en   252    50      0
#> 30                    Asthma follow-up      en   252    50      0
#> 31                    Suspected asthma      en   252    50      0
#> 32                          Sequoiosis      en   252    50      0
#> 33                    Colophony asthma      en   252    50      0
#> 34        Asthma monitoring check done      en   252    50      0
#> 35                    Asthma education      en   252    50      0
#> 36                    Intrinsic asthma      en   252    50      0
#> 37          Asthma night-time symptoms      en   252    50      0
#> 38                    Childhood asthma      en   252    50      0
#> 39                    Detergent asthma      en   252    50      0
#> 40                    Printers' asthma      en   252    50      0
#> 41                   Asthmatoid wheeze      en   252    50      0
#> 42                   Near fatal asthma      en   252    50      0
#> 43                 Occupational asthma      en   252    50      0
#> 44                   Occasional asthma      en   252    50      0
#> 45                   Late onset asthma      en   252    50      0
#> 46                   Factitious asthma      en   252    50      0
#> 47                   Asthma monitoring      en   252    50      0
#> 48              Flax-dressers' disease      en   252    50      0
#> 49                          Byssinosis      en   252    50      0
#> 50 Acute severe exacerbation of asthma      en   252    50      0
#>         searchAfter searchAfterArray
#> 1  WzcwODA5MDAwMl0=        708090002
#> 2  WzcwODA5MDAwMl0=        708090002
#> 3  WzcwODA5MDAwMl0=        708090002
#> 4  WzcwODA5MDAwMl0=        708090002
#> 5  WzcwODA5MDAwMl0=        708090002
#> 6  WzcwODA5MDAwMl0=        708090002
#> 7  WzcwODA5MDAwMl0=        708090002
#> 8  WzcwODA5MDAwMl0=        708090002
#> 9  WzcwODA5MDAwMl0=        708090002
#> 10 WzcwODA5MDAwMl0=        708090002
#> 11 WzcwODA5MDAwMl0=        708090002
#> 12 WzcwODA5MDAwMl0=        708090002
#> 13 WzcwODA5MDAwMl0=        708090002
#> 14 WzcwODA5MDAwMl0=        708090002
#> 15 WzcwODA5MDAwMl0=        708090002
#> 16 WzcwODA5MDAwMl0=        708090002
#> 17 WzcwODA5MDAwMl0=        708090002
#> 18 WzcwODA5MDAwMl0=        708090002
#> 19 WzcwODA5MDAwMl0=        708090002
#> 20 WzcwODA5MDAwMl0=        708090002
#> 21 WzcwODA5MDAwMl0=        708090002
#> 22 WzcwODA5MDAwMl0=        708090002
#> 23 WzcwODA5MDAwMl0=        708090002
#> 24 WzcwODA5MDAwMl0=        708090002
#> 25 WzcwODA5MDAwMl0=        708090002
#> 26 WzcwODA5MDAwMl0=        708090002
#> 27 WzcwODA5MDAwMl0=        708090002
#> 28 WzcwODA5MDAwMl0=        708090002
#> 29 WzcwODA5MDAwMl0=        708090002
#> 30 WzcwODA5MDAwMl0=        708090002
#> 31 WzcwODA5MDAwMl0=        708090002
#> 32 WzcwODA5MDAwMl0=        708090002
#> 33 WzcwODA5MDAwMl0=        708090002
#> 34 WzcwODA5MDAwMl0=        708090002
#> 35 WzcwODA5MDAwMl0=        708090002
#> 36 WzcwODA5MDAwMl0=        708090002
#> 37 WzcwODA5MDAwMl0=        708090002
#> 38 WzcwODA5MDAwMl0=        708090002
#> 39 WzcwODA5MDAwMl0=        708090002
#> 40 WzcwODA5MDAwMl0=        708090002
#> 41 WzcwODA5MDAwMl0=        708090002
#> 42 WzcwODA5MDAwMl0=        708090002
#> 43 WzcwODA5MDAwMl0=        708090002
#> 44 WzcwODA5MDAwMl0=        708090002
#> 45 WzcwODA5MDAwMl0=        708090002
#> 46 WzcwODA5MDAwMl0=        708090002
#> 47 WzcwODA5MDAwMl0=        708090002
#> 48 WzcwODA5MDAwMl0=        708090002
#> 49 WzcwODA5MDAwMl0=        708090002
#> 50 WzcwODA5MDAwMl0=        708090002

# Retrieve multiple concepts
concepts_find(conceptIds =  c("233604007", "68566005"))

#>   conceptId active definitionStatus           moduleId effectiveTime        id
#> 1  68566005   TRUE    FULLY_DEFINED 900000000000207008      20020131  68566005
#> 2 233604007   TRUE    FULLY_DEFINED 900000000000207008      20150131 233604007
#>                                      fsn.term fsn.lang
#> 1 Urinary tract infectious disease (disorder)       en
#> 2                        Pneumonia (disorder)       en
#>                            pt.term pt.lang total limit offset      searchAfter
#> 1 Urinary tract infectious disease      en     2    50      0 WzIzMzYwNDAwN10=
#> 2                        Pneumonia      en     2    50      0 WzIzMzYwNDAwN10=
#>   searchAfterArray
#> 1        233604007
#> 2        233604007

# Use the SNOMED-CT Expression Constraint Language
concepts_find(
  ecl = paste(
    "<! 68566005 | Urinary tract infectious disease (disorder) |",
    "AND",
    "< 87628006 | Bacterial infectious disease (disorder) |"
    )
  )

#>   conceptId active definitionStatus           moduleId effectiveTime        id
#> 1 312124009   TRUE    FULLY_DEFINED 900000000000207008      20030131 312124009
#>                                 fsn.term fsn.lang                     pt.term
#> 1 Bacterial urinary infection (disorder)       en Bacterial urinary infection
#>   pt.lang total limit offset      searchAfter searchAfterArray
#> 1      en     1    50      0 WzMxMjEyNDAwOV0=        312124009